Fixing errors in dataproviders
